+++ CHANGELOG.md
@@ -5,3 +5,6 @@
 
 ### 2018-09-25
 - Optimized log lines and log format
+
+### 2018-09-28
+- Build the grape cache in advance using `grape install`